% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneinfo}
\alias{geneinfo}
\title{C. elegans gene information}
\format{
The variables are as follows:
 \describe{
     \item WBGENEID. The Gene ID from WromBase
     \item SeqStart. Transcription starting site.
     \item CDS_Length. The largest CDS of this gene.
     \item CDS_ID. Transcription ID of the largest CDS.
     \item Gene_Length. The gene size includes intros.
     \item CHR. The chromosome that the gene locates.
     \item POS. Gene location in map unit.
     \item Gene_Name. the common name of the gene.
     \item PharynxEx. Is the gene expressed in pharynx? (TRUE or FALSE)
     \item Germlineex. Is the gene expressed in germline? (TRUE or FALSE)
     \item IntesEx. Is the gene expressed in intestine? (TRUE or FALSE)
}
}
\source{
{WormBase, WromMine, SimpleMine}
}
\usage{
geneinfo
}
\description{
A summary dataset containing all the C. elegans genes with its largest transcribe.
}
\examples{
data(geneinfo)    # Lazy loading
}
\keyword{datasets}
