% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hardmask_data.R
\docType{data}
\name{hardmask_list}
\alias{hardmask_list}
\title{C. elegans hardmask list}
\format{
The variables are as follows:
 \describe{
     \item CHR Chromosome.
     \item Seq Seuqence position.
}
}
\source{
{WormBase, Parasite}
}
\usage{
hardmask_list
}
\description{
A summary lsit containing all the C. elegans duplicated seuqeunce (hardmasked) postions based on PRJNA13758.WBPS18.
}
\examples{
data(hardmask_list)    # Lazy loading
}
\keyword{datasets}
