% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_fig.R
\name{mut_fig}
\alias{mut_fig}
\title{Generate a figure of  total mutations (left) and normalized mutation density (right)}
\usage{
mut_fig(
  input,
  chr,
  left_max,
  left_min,
  left_threshold,
  right_max,
  right_min,
  right_threshold
)
}
\arguments{
\item{1)}{Input the .txt file generated from pooled_analysis function, 2) Sepcifiy the chromosome, 3) Left max of the axis, 4) Left min of the axis, 5) Left threthold to display gene names, 6) Right min, 7) Right min, and 8) Right threshold.}
}
\description{
This function generate a figure including total mutations (left) and normalized mutation density (right).
}
\examples{
mut_fig()
}
\keyword{Mutation}
\keyword{density}
\keyword{fugure}
