% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmAF.R
\name{rmAF}
\alias{rmAF}
\title{Remove High Allele Frequency}
\usage{
rmAF(input, output, af)
}
\arguments{
\item{1)}{Input the raw .vcf files in a .txt format, 2) Name the output file, and 3) Designate the AF that you wanted to remove (from 0.0 to 1.0).}
}
\description{
This function does two things: 1) Separate a single position (nucleotide) with a maximum three different allele frequency reported from CRISP .vcf file into three different rows with the corresponding altered result and frequency, and 2) remove SNVs that have equal or greater than the designated AF.
}
\examples{
rmAF()
}
\keyword{Allele}
\keyword{Frequency}
